pyproject.toml
[project]
name = "rag-agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pyright>=1.1.408",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
    "uvicorn>=0.44.0",
]
